+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.11.0] - ???
+### Added
+- Live reporting of progress parsing files (below the component)
+
 ### Changed
 - Reduced minimum required Rhino 6 version to `6.10`